requirements.txt
# 新架构依赖
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
injector==0.21.0
asyncio
websockets==12.0
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1

# 保留原有依赖
ccxt==4.1.64
pandas==2.1.3
numpy>=1.26.0
aiohttp==3.9.1
websocket-client==1.6.4
pyyaml==6.0.1
rich>=13.0.0
hyperliquid-python-sdk
eth_account
